ENH: Add abstract integer types for use with image size
Add SizeValueType, IdentifierType, IndexValueType and OffsetValueType
as typedefs available in the top level itk namespace.

Use these instead of "long" or "unsigned long" to achieve 64-bit
image size capability on Win64.

Replace "long" and "unsigned long" with appropriate abstract types
everywhere in the code where they are presently used for one of the
above purposes. Other uses of long remain where their use does not
require a 64-bit value space.

+++ Code/BasicFilters/itkBinaryThresholdProjectionImageFilter.h
@@ -52,7 +52,7 @@ template< class TInputPixel, class TOutputPixel >
 class BinaryThresholdAccumulator
 {
 public:
-  BinaryThresholdAccumulator(unsigned long) {}
+  BinaryThresholdAccumulator(size_t) {}
   ~BinaryThresholdAccumulator(){}
 
   inline void Initialize()
@@ -177,7 +177,7 @@ protected:
        << std::endl;
   }
 
-  virtual AccumulatorType NewAccumulator(unsigned long size) const
+  virtual AccumulatorType NewAccumulator(size_t size) const
   {
     AccumulatorType accumulator(size);
